+++ colors.py
@@ -1,4 +1,4 @@
-# referrence: http://htmlcolorcodes.com/
+﻿# referrence: http://htmlcolorcodes.com/
 
 C_black = (0, 0, 0, 255)
 C_white = (255, 255, 255, 255)