+++ extend.py
@@ -1,4 +1,4 @@
-from expy import *
+﻿from expy import *
 np = shared.np
 
 '"Extend.py" is a play for any kind of plugin function'