+++ stim/display.py
@@ -1,4 +1,4 @@
-from expy import shared
+﻿from expy import shared
 from expy.colors import *
 from .draw import *
 from expy.response import *
@@ -19,6 +19,9 @@ def clear():
     # shared.pg.display.flip()
     shared.win.clear()
     shared.win.flip()
+    shared.win.clear()
+    shared.win.flip()
+
 
 def show(out_time=False, clean_screen=True, backup=None):
     '''